Updated Example app to latest React Native
