+++ Example/README.md
@@ -1,4 +1,6 @@
 ## react-native-admob example project
 
 1. `npm install`
-2. place **GoogleMobileAds.framework** from https://developers.google.com/admob/ios/download in ios folder
+2. `gem install cocoapods` (if not already installed)
+3. `(cd ios && pod install)`
+4. `react-native run-ios` or `react-native run-android`
