+++ Example/android/settings.gradle
@@ -1,5 +1,5 @@
 rootProject.name = 'Example'
-
-include ':app'
 include ':react-native-admob'
 project(':react-native-admob').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-admob/android')
+
+include ':app'
