+++ Example/app.json
@@ -1,0 +1,4 @@
+{
+  "name": "Example",
+  "displayName": "Example"
+}
