+++ Example/ios/Podfile
@@ -1,0 +1,13 @@
+# Uncomment the next line to define a global platform for your project
+# platform :ios, '9.0'
+
+target 'Example' do
+  # Uncomment the next line if you're using Swift or would like to use dynamic frameworks
+  # use_frameworks!
+
+  pod 'Firebase/AdMob'
+
+  # target 'sudoblockTests' do
+  #   inherit! :search_paths
+  # end
+end
